# apps/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: cc-tunnel API
  description: AI chat session management via claude CLI
  version: 1.0.0

paths:
  /conversations:
    post:
      operationId: createConversation
      summary: Create a new conversation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listConversations
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}:
    get:
      operationId: getConversation
      summary: Get conversation with message history
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteConversation
      summary: Delete a conversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /conversations/{conversationId}/messages:
    post:
      operationId: sendMessage
      summary: Send a message and stream the assistant response
      description: |
        Sends a user message to the conversation and streams the assistant
        response as Server-Sent Events (SSE).
        Response Content-Type: text/event-stream
        Each SSE event has format: data: <json>\n\n
        Event types: {"type":"text","content":"..."} | {"type":"done","session_id":"...","cost_usd":0.01} | {"type":"error","message":"..."}
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: SSE stream of assistant response
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
          default: ""
        model:
          type: string
          default: "claude-sonnet-4-6"
        system_prompt:
          type: string

    Conversation:
      type: object
      required:
        - id
        - title
        - model
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        model:
          type: string
        system_prompt:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Message:
      type: object
      required:
        - id
        - conversation_id
        - role
        - content
        - created_at
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time

    ConversationDetail:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          required:
            - messages
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"

    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: User message content

    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "ok"

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string